+++ src/Database/index.js
@@ -161,6 +161,8 @@ export default class Database {
 
   _resetCount: number = 0
 
+  _isBeingReset: boolean = false
+
   // Resets database - permanently destroys ALL records stored in the database, and sets up empty database
   //
   // NOTE: This is not 100% safe automatically and you must take some precautions to avoid bugs:
@@ -175,14 +177,37 @@ export default class Database {
     this._ensureInAction(
       `Database.unsafeResetDatabase() can only be called from inside of an Action. See docs for more details.`,
     )
-    // Doing this in very specific order:
-    // First kill actions, to ensure no more traffic to adapter happens
-    // then clear the database
-    // and only then clear caches, since might have had queued fetches from DB still bringing in items to cache
-    this._actionQueue._abortPendingActions()
-    await this.adapter.unsafeResetDatabase()
-    this._unsafeClearCaches()
-    this._resetCount += 1
+    try {
+      this._isBeingReset = true
+      // First kill actions, to ensure no more traffic to adapter happens
+      this._actionQueue._abortPendingActions()
+
+      // Kill ability to call adapter methods during reset (to catch bugs if someone does this)
+      const { adapter } = this
+      const ErrorAdapter = require('../adapters/error').default
+      this.adapter = (new ErrorAdapter(): any)
+
+      // Check for illegal subscribers
+      if (this._subscribers.length) {
+        // eslint-disable-next-line no-console
+        console.error(
+          `Application error! ${this._subscribers.length} Database subscriber was detected during database.unsafeResetDatabase() call. App should not hold onto subscriptions or Watermelon objects while resetting database.`,
+        )
+        this._subscribers = []
+      }
+
+      // Clear the database
+      await adapter.unsafeResetDatabase()
+
+      // Only now clear caches, since there may have been queued fetches from DB still bringing in items to cache
+      this._unsafeClearCaches()
+
+      // Restore working Database
+      this._resetCount += 1
+      this.adapter = adapter
+    } finally {
+      this._isBeingReset = false
+    }
   }
 
   _unsafeClearCaches(): void {