+++ CHANGELOG.md
@@ -14,6 +14,7 @@ All notable changes to this project will be documented in this file.
 - [Typescript] Fixed types of decorators.
 - [Typescript] Add Tests to test Types.
 - Fixed typo in learn-to-use docs.
+- [Typescript] Fixed types of changes.
 
 ## 0.16 - 2020-03-06