+++ CHANGELOG.md
@@ -18,6 +18,7 @@ All notable changes to this project will be documented in this file.
 - [Typescript] Add Tests to test Types.
 - Fixed typo in learn-to-use docs.
 - [Typescript] Fixed types of changes.
+- [Typescript] Fixed types of Relation
 
 ## 0.16 - 2020-03-06